# secp256k1: Optimize pubkey parse.