**This requires #2163**.

This modifies the `ParsePubKey` function to avoid an additional unnecessary check if the point is on the curve in the compressed pubkey case because it necessarily has already determined if that is true via the point decompression from the x coordinate.

Also, prevent x from escaping to the heap for the error print in compressed case while here.

```
benchmark                          old ns/op    new ns/op   delta
------------------------------------------------------------------
BenchmarkParsePubKeyCompressed     11901        11745       -1.31%
BenchmarkParsePubKeyUncompressed   335          325         -2.99%

benchmark                          old allocs   new allocs  delta
-------------------------------------------------------------------
BenchmarkParsePubKeyCompressed     2            1           -50.00%
BenchmarkParsePubKeyUncompressed   2            1           -50.00%

benchmark                          old bytes    new bytes   delta
-------------------------------------------------------------------
BenchmarkParsePubKeyCompressed     128          80          -37.50%
BenchmarkParsePubKeyUncompressed   128          80          -37.50%
```